Complete overhaul of the transposition table and setting up the future foundation for Lazy SMP

+++ src/types/constants.rs
@@ -32,5 +32,5 @@ pub const MATE_SCORE: i32 = 9000;
 pub const MATE_CUTOFF: i32 = 8900;
 pub const TIMEOUT_SCORE: i32 = 111111;
 
-pub const MAX_DEPTH: usize = 128;
+pub const MAX_DEPTH: u8 = 128;
 pub const MAX_HISTORY: i32 = 8000;
